--- src/axios-auth.js
@@ -1,8 +1,0 @@
-import axios from 'axios'
-
-const instance = axios.create({
-  baseURL:
-    'https://identitytoolkit.googleapis.com/v1'
-});
-
-export default instance
